+++ packages/native/tests/types/types.test.ts
@@ -48,6 +48,7 @@ recipe.createPage(595, 842, { left: 36 }).margins({ top: 36 });
 var pageBox: muhammara.PDFPageBoxType = muhammara.ePDFPageBoxCropBox;
 recipe.setPageBox(pageBox, 10, 20, 585, 822);
 recipe.setPageBox(muhammara.ePDFPageBoxMediaBox, 0, 0, 595, 842);
+recipe.rotate(90).endPage();
 var margins: Required<muhammara.Recipe.RecipeMargins> = recipe.margins();
 var title: string = recipe.getPageInfo().title;
 var textWidth: number = recipe.textDimensions("text").width;